Add authz test model and remove placeholder UUIDs
