+++ tests/integration/identity/oauth2/authz/authz_test.go
@@ -34,11 +34,10 @@ import (
 )
 
 const (
-	clientID      = "authz_test_client_123"
-	clientSecret  = "authz_test_secret_123"
-	appName       = "AuthzTestApp"
-	redirectURI   = "https://localhost:3000"
-	testServerURL = "https://localhost:8095"
+	clientID     = "authz_test_client_123"
+	clientSecret = "authz_test_secret_123"
+	appName      = "AuthzTestApp"
+	redirectURI  = "https://localhost:3000"
 )
 
 var (
@@ -303,7 +302,7 @@ func (ts *AuthzTestSuite) TestBasicAuthorizationRequest() {
 
 	for _, tc := range testCases {
 		ts.Run(tc.Name, func() {
-			resp, err := initiateAuthorizationFlow(testServerURL, tc.ClientID, tc.RedirectURI, tc.ResponseType, "openid", tc.State)
+			resp, err := initiateAuthorizationFlow(tc.ClientID, tc.RedirectURI, tc.ResponseType, "openid", tc.State)
 			ts.NoError(err, "Failed to initiate authorization flow")
 			defer resp.Body.Close()
 
@@ -334,16 +333,16 @@ func (ts *AuthzTestSuite) TestTokenRequestValidation() {
 	username := "token_test_user"
 	password := "testpass123"
 
-	userID, err := createTestUser(testServerURL, username, password, testOUID)
+	userID, err := createTestUser(username, password, testOUID)
 	ts.NoError(err, "Failed to create test user")
 	defer func() {
-		if err := deleteTestUser(testServerURL, userID); err != nil {
+		if err := deleteTestUser(userID); err != nil {
 			ts.T().Logf("Warning: Failed to delete test user: %v", err)
 		}
 	}()
 
 	// Get a valid authorization code first
-	resp, err := initiateAuthorizationFlow(testServerURL, clientID, redirectURI, "code", "openid", "token_test_state")
+	resp, err := initiateAuthorizationFlow(clientID, redirectURI, "code", "openid", "token_test_state")
 	ts.NoError(err, "Failed to initiate authorization flow")
 	defer resp.Body.Close()
 
@@ -353,7 +352,7 @@ func (ts *AuthzTestSuite) TestTokenRequestValidation() {
 	ts.NoError(err, "Failed to extract session data")
 
 	// Execute authentication flow
-	flowStep, err := ExecuteAuthenticationFlow(testServerURL, ts.applicationID, map[string]string{
+	flowStep, err := ExecuteAuthenticationFlow(ts.applicationID, map[string]string{
 		"username": username,
 		"password": password,
 	})
@@ -361,7 +360,7 @@ func (ts *AuthzTestSuite) TestTokenRequestValidation() {
 	ts.Equal("COMPLETE", flowStep.FlowStatus, "Flow should complete successfully")
 
 	// Complete authorization
-	authzResponse, err := completeAuthorization(testServerURL, sessionDataKey, flowStep.Assertion)
+	authzResponse, err := completeAuthorization(sessionDataKey, flowStep.Assertion)
 	ts.NoError(err, "Failed to complete authorization")
 	validAuthzCode, err := extractAuthorizationCode(authzResponse.RedirectURI)
 	ts.NoError(err, "Failed to extract authorization code")
@@ -500,7 +499,7 @@ func (ts *AuthzTestSuite) TestTokenRequestValidation() {
 
 	for _, tc := range testCases {
 		ts.Run(tc.Name, func() {
-			result, err := requestToken(testServerURL, tc.ClientID, tc.ClientSecret, tc.Code, tc.RedirectURI, tc.GrantType)
+			result, err := requestToken(tc.ClientID, tc.ClientSecret, tc.Code, tc.RedirectURI, tc.GrantType)
 			ts.NoError(err, "Token request should not error at transport level")
 			ts.Equal(tc.ExpectedStatus, result.StatusCode, "Expected status code")
 
@@ -607,7 +606,7 @@ func (ts *AuthzTestSuite) TestRedirectURIValidation() {
 
 	for _, tc := range testCases {
 		ts.Run(tc.Name, func() {
-			resp, err := initiateAuthorizationFlow(testServerURL, tc.ClientID, tc.RedirectURI, tc.ResponseType, tc.Scope, tc.State)
+			resp, err := initiateAuthorizationFlow(tc.ClientID, tc.RedirectURI, tc.ResponseType, tc.Scope, tc.State)
 			ts.NoError(err, "Failed to initiate authorization flow")
 			defer resp.Body.Close()
 
@@ -660,7 +659,7 @@ func (ts *AuthzTestSuite) TestCompleteAuthorizationCodeFlow() {
 	for _, tc := range testCases {
 		ts.Run(tc.Name, func() {
 			// Create test user with credentials
-			userID, err := createTestUser(testServerURL, tc.Username, tc.Password, testOUID)
+			userID, err := createTestUser(tc.Username, tc.Password, testOUID)
 			if err != nil {
 				ts.T().Fatalf("Failed to create test user: %v", err)
 			}
@@ -669,13 +668,13 @@ func (ts *AuthzTestSuite) TestCompleteAuthorizationCodeFlow() {
 			}
 
 			defer func() {
-				if err := deleteTestUser(testServerURL, userID); err != nil {
+				if err := deleteTestUser(userID); err != nil {
 					ts.T().Logf("Warning: Failed to delete test user: %v", err)
 				}
 			}()
 
 			// Start authorization flow
-			resp, err := initiateAuthorizationFlow(testServerURL, tc.ClientID, tc.RedirectURI, tc.ResponseType, tc.Scope, tc.State)
+			resp, err := initiateAuthorizationFlow(tc.ClientID, tc.RedirectURI, tc.ResponseType, tc.Scope, tc.State)
 			ts.NoError(err, "Failed to initiate authorization flow")
 			defer resp.Body.Close()
 
@@ -694,7 +693,7 @@ func (ts *AuthzTestSuite) TestCompleteAuthorizationCodeFlow() {
 			}
 
 			// Execute authentication flow
-			flowStep, err := ExecuteAuthenticationFlow(testServerURL, ts.applicationID, map[string]string{
+			flowStep, err := ExecuteAuthenticationFlow(ts.applicationID, map[string]string{
 				"username": tc.Username,
 				"password": tc.Password,
 			})
@@ -717,7 +716,7 @@ func (ts *AuthzTestSuite) TestCompleteAuthorizationCodeFlow() {
 			}
 
 			// Complete authorization
-			authzResponse, err := completeAuthorization(testServerURL, sessionDataKey, flowStep.Assertion)
+			authzResponse, err := completeAuthorization(sessionDataKey, flowStep.Assertion)
 			ts.NoError(err, "Failed to complete authorization")
 			ts.NotEmpty(authzResponse.RedirectURI, "Redirect URI should be present")
 
@@ -726,7 +725,7 @@ func (ts *AuthzTestSuite) TestCompleteAuthorizationCodeFlow() {
 			ts.NotEmpty(authzCode, "Authorization code should be present")
 
 			// Exchange authorization code for access token
-			result, err := requestToken(testServerURL, clientID, clientSecret, authzCode, tc.RedirectURI, "authorization_code")
+			result, err := requestToken(clientID, clientSecret, authzCode, tc.RedirectURI, "authorization_code")
 			ts.NoError(err, "Failed to exchange code for token")
 			ts.Equal(http.StatusOK, result.StatusCode, "Token request should succeed")
 			tokenResponse := result.Token
@@ -771,16 +770,16 @@ func (ts *AuthzTestSuite) TestAuthorizationCodeErrorScenarios() {
 	for _, tc := range testCases {
 		ts.Run(tc.Name, func() {
 			// Create test user
-			userID, err := createTestUser(testServerURL, tc.Username, tc.Password, testOUID)
+			userID, err := createTestUser(tc.Username, tc.Password, testOUID)
 			ts.NoError(err, "Failed to create test user")
 			defer func() {
-				if err := deleteTestUser(testServerURL, userID); err != nil {
+				if err := deleteTestUser(userID); err != nil {
 					ts.T().Logf("Warning: Failed to delete test user: %v", err)
 				}
 			}()
 
 			// Start authorization flow
-			resp, err := initiateAuthorizationFlow(testServerURL, tc.ClientID, tc.RedirectURI, tc.ResponseType, tc.Scope, tc.State)
+			resp, err := initiateAuthorizationFlow(tc.ClientID, tc.RedirectURI, tc.ResponseType, tc.Scope, tc.State)
 			ts.NoError(err, "Failed to initiate authorization flow")
 			defer resp.Body.Close()
 
@@ -793,7 +792,7 @@ func (ts *AuthzTestSuite) TestAuthorizationCodeErrorScenarios() {
 			ts.NoError(err, "Failed to extract session data")
 
 			// Execute authentication flow
-			flowStep, err := ExecuteAuthenticationFlow(testServerURL, ts.applicationID, map[string]string{
+			flowStep, err := ExecuteAuthenticationFlow(ts.applicationID, map[string]string{
 				"username": tc.Username,
 				"password": tc.Password,
 			})
@@ -804,7 +803,7 @@ func (ts *AuthzTestSuite) TestAuthorizationCodeErrorScenarios() {
 				ts.T().Fatalf("Expected flow status COMPLETE, got %s", flowStep.FlowStatus)
 			}
 
-			authzResponse, err := completeAuthorization(testServerURL, sessionDataKey, flowStep.Assertion)
+			authzResponse, err := completeAuthorization(sessionDataKey, flowStep.Assertion)
 			if err != nil {
 				ts.T().Fatalf("Failed to complete authorization: %v", err)
 			}
@@ -814,12 +813,12 @@ func (ts *AuthzTestSuite) TestAuthorizationCodeErrorScenarios() {
 			ts.NoError(err, "Failed to extract authorization code")
 
 			if tc.Name == "Reused Authorization Code" {
-				result, err := requestToken(testServerURL, clientID, clientSecret, authzCode, tc.RedirectURI, "authorization_code")
+				result, err := requestToken(clientID, clientSecret, authzCode, tc.RedirectURI, "authorization_code")
 				ts.NoError(err, "First token exchange should succeed")
 				ts.Equal(http.StatusOK, result.StatusCode, "First token exchange should succeed")
 
 				// Second attempt should fail with invalid_grant
-				result2, err := requestToken(testServerURL, clientID, clientSecret, authzCode, tc.RedirectURI, "authorization_code")
+				result2, err := requestToken(clientID, clientSecret, authzCode, tc.RedirectURI, "authorization_code")
 				ts.NoError(err, "Second token exchange should not error at transport level")
 				ts.Equal(http.StatusBadRequest, result2.StatusCode, "Second token exchange should fail with bad request")
