+++ tests/integration/identity/oauth2/authz/utils.go
@@ -28,82 +28,12 @@ import (
 	"net/url"
 )
 
-// TestCase represents a test case for authorization code flow
-type TestCase struct {
-	Name           string
-	ClientID       string
-	RedirectURI    string
-	ResponseType   string
-	Scope          string
-	State          string
-	Username       string
-	Password       string
-	ExpectedStatus int
-	ExpectedError  string
-}
-
-// User represents a test user for authorization code tests
-type User struct {
-	OrganizationUnit string                 `json:"organizationUnit"`
-	Type             string                 `json:"type"`
-	Attributes       map[string]interface{} `json:"attributes"`
-}
-
-// FlowResponse represents the response from flow execution
-type FlowResponse struct {
-	FlowID        string    `json:"flowId"`
-	FlowStatus    string    `json:"flowStatus"`
-	Type          string    `json:"type"`
-	Data          *FlowData `json:"data,omitempty"`
-	Assertion     string    `json:"assertion,omitempty"`
-	FailureReason string    `json:"failureReason,omitempty"`
-}
-
-// FlowData represents the data returned by flow execution
-type FlowData struct {
-	Inputs []FlowInput `json:"inputs,omitempty"`
-}
-
-// FlowInput represents an input required by the flow
-type FlowInput struct {
-	Name     string `json:"name"`
-	Type     string `json:"type"`
-	Required bool   `json:"required"`
-}
-
-// AuthorizationResponse represents the response from authorization completion
-type AuthorizationResponse struct {
-	RedirectURI string `json:"redirect_uri"`
-}
-
-// TokenResponse represents the response from token exchange
-type TokenResponse struct {
-	AccessToken  string  `json:"access_token"`
-	TokenType    string  `json:"token_type"`
-	ExpiresIn    float64 `json:"expires_in"`
-	Scope        string  `json:"scope,omitempty"`
-	RefreshToken string  `json:"refresh_token,omitempty"`
-}
-
-// TokenHTTPResult captures raw HTTP response details from the token endpoint.
-type TokenHTTPResult struct {
-	StatusCode int
-	Body       []byte
-	Token      *TokenResponse
-}
-
-// FlowStep represents a single step in a flow execution
-type FlowStep struct {
-	FlowID        string    `json:"flowId"`
-	FlowStatus    string    `json:"flowStatus"`
-	Type          string    `json:"type"`
-	Data          *FlowData `json:"data,omitempty"`
-	Assertion     string    `json:"assertion,omitempty"`
-	FailureReason string    `json:"failureReason,omitempty"`
-}
+const (
+	testServerURL = "https://localhost:8095"
+)
 
 // createTestUser creates a test user with the given credentials
-func createTestUser(testServerURL string, username, password, ouID string) (string, error) {
+func createTestUser(username, password, ouID string) (string, error) {
 	userData := User{
 		OrganizationUnit: ouID,
 		Type:             "person",
@@ -159,7 +89,7 @@ func createTestUser(testServerURL string, username, password, ouID string) (stri
 }
 
 // deleteTestUser deletes a test user by ID
-func deleteTestUser(testServerURL string, userID string) error {
+func deleteTestUser(userID string) error {
 	req, err := http.NewRequest("DELETE", testServerURL+"/users/"+userID, nil)
 	if err != nil {
 		return fmt.Errorf("failed to create delete request: %w", err)
@@ -185,7 +115,7 @@ func deleteTestUser(testServerURL string, userID string) error {
 }
 
 // initiateAuthorizationFlow starts the OAuth2 authorization flow
-func initiateAuthorizationFlow(testServerURL string, clientID, redirectURI, responseType, scope, state string) (*http.Response, error) {
+func initiateAuthorizationFlow(clientID, redirectURI, responseType, scope, state string) (*http.Response, error) {
 	authURL := testServerURL + "/oauth2/authorize"
 	params := url.Values{}
 	params.Set("client_id", clientID)
@@ -217,7 +147,7 @@ func initiateAuthorizationFlow(testServerURL string, clientID, redirectURI, resp
 }
 
 // ExecuteAuthenticationFlow executes an authentication flow and returns the flow step
-func ExecuteAuthenticationFlow(testServerURL string, applicationId string, inputs map[string]string) (*FlowStep, error) {
+func ExecuteAuthenticationFlow(applicationId string, inputs map[string]string) (*FlowStep, error) {
 	flowData := map[string]interface{}{
 		"applicationId": applicationId,
 		"flowType":      "AUTHENTICATION",
@@ -264,7 +194,7 @@ func ExecuteAuthenticationFlow(testServerURL string, applicationId string, input
 }
 
 // completeAuthorization completes the authorization using the assertion
-func completeAuthorization(testServerURL string, sessionDataKey, assertion string) (*AuthorizationResponse, error) {
+func completeAuthorization(sessionDataKey, assertion string) (*AuthorizationResponse, error) {
 	authzData := map[string]interface{}{
 		"sessionDataKey": sessionDataKey,
 		"assertion":      assertion,
@@ -309,7 +239,7 @@ func completeAuthorization(testServerURL string, sessionDataKey, assertion strin
 
 // requestToken performs a token request and returns raw HTTP result for both success and failure scenarios.
 // grantType, code, and redirectURI are sent in the form body, while client credentials are sent via HTTP Basic Auth header.
-func requestToken(testServerURL string, clientID, clientSecret, code, redirectURI, grantType string) (*TokenHTTPResult, error) {
+func requestToken(clientID, clientSecret, code, redirectURI, grantType string) (*TokenHTTPResult, error) {
 	tokenURL := testServerURL + "/oauth2/token"
 	tokenData := url.Values{}
 
@@ -440,20 +370,6 @@ func validateOAuth2ErrorRedirect(location string, expectedError string, expected
 	}
 
 	return fmt.Errorf("no error parameters found in redirect URL (neither 'error'/'error_description' nor 'errorCode'/'errorMessage')")
-}
-
-// IDP structures for OAuth2 tests
-type IDPProperty struct {
-	Name     string `json:"name"`
-	Value    string `json:"value"`
-	IsSecret bool   `json:"is_secret"`
-}
-
-type IDP struct {
-	ID          string        `json:"id,omitempty"`
-	Name        string        `json:"name"`
-	Description string        `json:"description"`
-	Properties  []IDPProperty `json:"properties"`
 }
 
 // createIdp creates an identity provider via API
